move common function to a util class
